Extract Formatting to FunctionDocFormatter, simplify showUserFunctions
