+++ src/net/sourceforge/kolmafia/KoLmafiaASH.java
@@ -1,7 +1,6 @@
 package net.sourceforge.kolmafia;
 
 import java.io.File;
-import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.Iterator;
 import java.util.List;
@@ -15,8 +14,6 @@ import net.sourceforge.kolmafia.textui.Parser;
 import net.sourceforge.kolmafia.textui.ScriptRuntime;
 import net.sourceforge.kolmafia.textui.javascript.JavascriptRuntime;
 import net.sourceforge.kolmafia.textui.parsetree.Function;
-import net.sourceforge.kolmafia.textui.parsetree.FunctionList;
-import net.sourceforge.kolmafia.textui.parsetree.LibraryFunction;
 import net.sourceforge.kolmafia.textui.parsetree.VariableReference;
 import net.sourceforge.kolmafia.utilities.FileUtilities;
 
@@ -243,124 +240,38 @@ public abstract class KoLmafiaASH {
   }
 
   public static void showUserFunctions(final AshRuntime interpreter, final String filter) {
-    Formatting.showFunctions(interpreter.getFunctions(), filter.toLowerCase(), false);
-  }
+    var functions = interpreter.getFunctions();
+    var lowerFilter = filter.toLowerCase();
 
-  public static class Formatting {
-    private static final String DOC_COLOR_LIGHT = "#666666";
-    private static final String DOC_COLOR_DARK = "#888888";
-
-    private Formatting() {}
-
-    private static String docColor() {
-      return KoLmafiaGUI.isDarkTheme() ? DOC_COLOR_DARK : DOC_COLOR_LIGHT;
+    if (functions.isEmpty()) {
+      RequestLogger.printLine("No functions in your current namespace.");
+      return;
     }
 
-    public static String formatDocBlock(LibraryFunction func) {
-      var description = func.getDescription();
-      var params = func.getVariableReferences();
-
-      var lines = new ArrayList<String>();
-
-      if (description != null && !description.isEmpty()) {
-        lines.add(description);
-      }
-
-      for (VariableReference var : params) {
-        var paramDescription = var.getDescription();
-        if (paramDescription != null && !paramDescription.isEmpty()) {
-          lines.add("@param " + var.getName() + " " + paramDescription);
+    for (Function func : functions) {
+      boolean matches = lowerFilter.isEmpty();
+      if (!matches) matches = func.getName().toLowerCase().contains(lowerFilter);
+      if (!matches) {
+        for (VariableReference ref : func.getVariableReferences()) {
+          String refType = ref.getType().toString();
+          matches |= refType != null && refType.contains(lowerFilter);
         }
       }
+      if (!matches) continue;
 
-      if (lines.isEmpty()) {
-        return null;
+      var signature = new StringBuilder();
+      signature.append(func.getType()).append(" ").append(func.getName()).append("( ");
+      String sep = "";
+      for (VariableReference var : func.getVariableReferences()) {
+        signature.append(sep);
+        sep = ", ";
+        signature.append(var.getRawType());
+        if (var.getName() != null) {
+          signature.append(" ").append(var.getName());
+        }
       }
-
-      var color = docColor();
-      var sb = new StringBuilder();
-      sb.append("<font color='").append(color).append("'>");
-
-      if (lines.size() == 1) {
-        sb.append("/** ").append(lines.get(0)).append(" */");
-      } else {
-        sb.append("/**");
-        for (var line : lines) {
-          sb.append("<br>&nbsp;* ").append(line);
-        }
-        sb.append("<br>&nbsp;*/");
-      }
-
-      sb.append("</font>");
-      return sb.toString();
-    }
-
-    public static void showFunctions(
-        final FunctionList functions, final String filter, boolean addLinks) {
-      addLinks = addLinks && StaticEntity.isGUIRequired();
-
-      if (functions.isEmpty()) {
-        RequestLogger.printLine("No functions in your current namespace.");
-        return;
-      }
-
-      for (Function func : functions) {
-        boolean matches = filter.isEmpty();
-
-        if (!matches) {
-          matches = func.getName().toLowerCase().contains(filter);
-        }
-
-        if (!matches) {
-          for (VariableReference ref : func.getVariableReferences()) {
-            String refType = ref.getType().toString();
-            matches |= refType != null && refType.contains(filter);
-          }
-        }
-
-        if (!matches) {
-          continue;
-        }
-
-        if (func instanceof LibraryFunction lf) {
-          var docBlock = formatDocBlock(lf);
-          if (docBlock != null) {
-            RequestLogger.printHtml(docBlock);
-          }
-        }
-
-        StringBuilder signature = new StringBuilder();
-
-        signature.append(func.getType());
-        signature.append(" ");
-        if (addLinks) {
-          signature.append("<a href='https://wiki.kolmafia.us/index.php?title=");
-          signature.append(func.getName());
-          signature.append("'>");
-        }
-        signature.append(func.getName());
-        if (addLinks) {
-          signature.append("</a>");
-        }
-        signature.append("( ");
-
-        String sep = "";
-        for (VariableReference var : func.getVariableReferences()) {
-          signature.append(sep);
-          sep = ", ";
-
-          signature.append(var.getRawType());
-
-          if (var.getName() != null) {
-            signature.append(" ");
-            signature.append(var.getName());
-          }
-        }
-
-        signature.append(" )");
-
-        RequestLogger.printHtml(signature.toString());
-      }
+      signature.append(" )");
+      RequestLogger.printHtml(signature.toString());
     }
   }
