+++ src/net/sourceforge/kolmafia/textui/command/AshRefCommand.java
@@ -1,6 +1,5 @@
 package net.sourceforge.kolmafia.textui.command;
 
-import net.sourceforge.kolmafia.KoLmafiaASH;
 import net.sourceforge.kolmafia.textui.RuntimeLibrary;
 
 public class AshRefCommand extends AbstractCommand {
@@ -10,7 +9,7 @@ public class AshRefCommand extends AbstractCommand {
 
   @Override
   public void run(final String cmd, final String parameters) {
-    KoLmafiaASH.Formatting.showFunctions(
+    FunctionDocFormatter.showFunctions(
         RuntimeLibrary.getFunctions(), parameters.toLowerCase(), true);
   }
 }
